Refactor and re-enable recomposer tests
Moved some of the recompose tests to unit tests from
integration tests. Removed tests that are redundant with
other composition tests.

Fixes: b/179279455
Test: ./gradlew :compose:r:r:i-t:cC
