+++ compose/runtime/runtime/integration-tests/src/androidAndroidTest/kotlin/androidx/compose/runtime/EmittableComposer.kt
@@ -16,7 +16,10 @@
 package androidx.compose.runtime
 
 import android.view.View
+import android.widget.Button
+import android.widget.TextView
 import android.widget.LinearLayout
+import androidx.compose.ui.viewinterop.AndroidView
 
 @Suppress("UNUSED_PARAMETER")
 @Composable
@@ -24,7 +27,14 @@ fun TextView(
     id: Int = 0,
     text: String = "",
     onClickListener: View.OnClickListener? = null
-) {}
+) {
+    AndroidView(viewBlock = { TextView(it) }) { view ->
+        view.id = id
+        view.text = text
+        if (onClickListener != null)
+            view.setOnClickListener(onClickListener)
+    }
+}
 
 @Suppress("UNUSED_PARAMETER")
 @Composable
@@ -32,7 +42,14 @@ fun Button(
     id: Int = 0,
     text: String = "",
     onClickListener: View.OnClickListener? = null
-) {}
+) {
+    AndroidView(viewBlock = { Button(it) }) { view ->
+        view.id = id
+        view.text = text
+        if (onClickListener != null)
+            view.setOnClickListener(onClickListener)
+    }
+}
 
 @Suppress("UNUSED_PARAMETER")
 @Composable
