+++ compose/runtime/runtime/integration-tests/src/androidAndroidTest/kotlin/androidx/compose/runtime/RecomposerTests.kt
@@ -17,8 +17,6 @@
 package androidx.compose.runtime
 
 import android.view.View
-import android.view.ViewGroup
-import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.compose.runtime.snapshots.Snapshot
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -32,11 +30,9 @@ import kotlinx.coroutines.cancelAndJoin
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.test.runBlockingTest
-import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import kotlin.test.assertNotNull
 
 @MediumTest
 @RunWith(AndroidJUnit4::class)
@@ -46,7 +42,6 @@ class RecomposerTests : BaseComposeTest() {
     override val activityRule = makeTestActivityRule()
 
     @Test
-    @Ignore("b/179279455")
     fun testNativeViewWithAttributes() {
         compose {
             TextView(id = 456, text = "some text")
@@ -61,7 +56,6 @@ class RecomposerTests : BaseComposeTest() {
     }
 
     @Test
-    @Ignore("b/179279455")
     fun testSlotKeyChangeCausesRecreate() {
         var i = 1
         var tv1: TextView? = null
@@ -95,158 +89,6 @@ class RecomposerTests : BaseComposeTest() {
         }
     }
 
-    @Test
-    @Ignore("b/179279455")
-    fun testViewWithViewChildren() {
-        compose {
-            LinearLayout(id = 345) {
-                TextView(id = 456, text = "some text")
-                TextView(id = 567, text = "some text")
-            }
-        }.then { activity ->
-            val ll = activity.findViewById(345) as LinearLayout
-            val tv1 = activity.findViewById(456) as TextView
-            val tv2 = activity.findViewById(567) as TextView
-
-            assertEquals(
-                "The linear layout should be the only child of root", 1,
-                activity.root.childCount
-            )
-            assertEquals("Both children should have been added", 2, ll.childCount)
-            assertTrue(
-                "Should be the expected TextView (1)",
-                ll.getChildAt(0) === tv1
-            )
-            assertTrue(
-                "Should be the expected TextView (2)",
-                ll.getChildAt(1) === tv2
-            )
-        }
-    }
-
-    @Test
-    @Ignore("b/179279455")
-    fun testForLoop() {
-        val items = listOf(1, 2, 3, 4, 5, 6)
-        compose {
-            LinearLayout(id = 345) {
-                for (i in items) {
-                    TextView(id = 456, text = "some text $i")
-                }
-            }
-        }.then { activity ->
-            val ll = activity.findViewById(345) as LinearLayout
-
-            assertEquals(
-                "The linear layout should be the only child of root", 1,
-                activity.root.childCount
-            )
-            assertEquals("Each item in the for loop should be a child", items.size, ll.childCount)
-            items.forEachIndexed { index, i ->
-                assertEquals(
-                    "Should be the correct child", "some text $i",
-                    (ll.getChildAt(index) as TextView).text
-                )
-            }
-        }
-    }
-
-    @Test
-    @Ignore("b/179279455")
-    fun testRecompose() {
-        val counter = Counter()
-
-        compose {
-            RecomposeTestComponentsA(
-                counter,
-                ClickAction.Recompose
-            )
-        }.then { activity ->
-            // everything got rendered once
-            assertEquals(1, counter["A"])
-            assertEquals(1, counter["100"])
-            assertEquals(1, counter["101"])
-            assertEquals(1, counter["102"])
-
-            (activity.findViewById(100) as TextView).performClick()
-            (activity.findViewById(102) as TextView).performClick()
-
-            // nothing should happen synchronously
-            assertEquals(1, counter["A"])
-            assertEquals(1, counter["100"])
-            assertEquals(1, counter["101"])
-            assertEquals(1, counter["102"])
-        }.then { activity ->
-            // only the clicked view got rerendered
-            assertEquals(1, counter["A"])
-            assertEquals(2, counter["100"])
-            assertEquals(1, counter["101"])
-            assertEquals(2, counter["102"])
-
-            // recompose() both the parent and the child... and show that the child only
-            // recomposes once as a result
-            (activity.findViewById(99) as LinearLayout).performClick()
-            (activity.findViewById(102) as TextView).performClick()
-        }.then {
-
-            assertEquals(2, counter["A"])
-            assertEquals(3, counter["100"])
-            assertEquals(2, counter["101"])
-            assertEquals(3, counter["102"])
-        }
-    }
-
-    @Test
-    @Ignore("b/179279455")
-    fun testRootRecompose() {
-        val counter = Counter()
-        val trigger = Trigger()
-
-        val listener =
-            ClickAction.PerformOnView {
-                trigger.recompose()
-            }
-
-        compose {
-            trigger.subscribe()
-            RecomposeTestComponentsA(
-                counter,
-                listener
-            )
-        }.then { activity ->
-            // everything got rendered once
-            assertEquals(1, counter["A"])
-            assertEquals(1, counter["100"])
-            assertEquals(1, counter["101"])
-            assertEquals(1, counter["102"])
-
-            (activity.findViewById(100) as TextView).performClick()
-            (activity.findViewById(102) as TextView).performClick()
-
-            // nothing should happen synchronously
-            assertEquals(1, counter["A"])
-            assertEquals(1, counter["100"])
-            assertEquals(1, counter["101"])
-            assertEquals(1, counter["102"])
-        }.then { activity ->
-            // as we recompose ROOT on every tap, everything should be increased once, because two
-            // clicks layed to one frame. None of these components are skippable, so each increments
-            assertEquals(2, counter["A"])
-            assertEquals(2, counter["100"])
-            assertEquals(2, counter["101"])
-            assertEquals(2, counter["102"])
-
-            (activity.findViewById(99) as LinearLayout).performClick()
-            (activity.findViewById(102) as TextView).performClick()
-        }.then {
-            // again, no matter what we tapped, we want to recompose root, so all counts increased
-            assertEquals(3, counter["A"])
-            assertEquals(3, counter["100"])
-            assertEquals(3, counter["101"])
-            assertEquals(3, counter["102"])
-        }
-    }
-
     // components for testing recompose behavior above
     sealed class ClickAction {
         object Recompose : ClickAction()
@@ -260,7 +102,7 @@ class RecomposerTests : BaseComposeTest() {
 
         TextView(
             id = id,
-            onClickListener = View.OnClickListener {
+            onClickListener = {
                 @Suppress("DEPRECATION")
                 when (listener) {
                     is ClickAction.Recompose -> scope.invalidate()
@@ -268,208 +110,6 @@ class RecomposerTests : BaseComposeTest() {
                 }
             }
         )
-    }
-
-    @Composable fun RecomposeTestComponentsA(counter: Counter, listener: ClickAction) {
-        counter.inc("A")
-        val scope = currentRecomposeScope
-        LinearLayout(
-            id = 99,
-            onClickListener = View.OnClickListener {
-                @Suppress("DEPRECATION")
-                when (listener) {
-                    is ClickAction.Recompose -> scope.invalidate()
-                    is ClickAction.PerformOnView -> listener.action.invoke(it)
-                }
-            }
-        ) {
-            for (id in 100..102) {
-                key(id) {
-                    RecomposeTestComponentsB(
-                        counter,
-                        listener,
-                        id
-                    )
-                }
-            }
-        }
-    }
-
-    @Test
-    @Ignore("b/179279455")
-    fun testCorrectViewTree() {
-        compose {
-            LinearLayout {
-                LinearLayout { }
-                LinearLayout { }
-            }
-            LinearLayout { }
-        }.then { activity ->
-            assertChildHierarchy(activity.root.viewBlockHolders()) {
-                """
-                    <LinearLayout>
-                        <LinearLayout />
-                        <LinearLayout />
-                    </LinearLayout>
-                    <LinearLayout />
-                """
-            }
-        }
-    }
-
-    @Test
-    @Ignore("b/179279455")
-    fun testCorrectViewTreeWithComponents() {
-
-        @Composable fun B() {
-            TextView()
-        }
-
-        compose {
-            LinearLayout {
-                LinearLayout {
-                    B()
-                }
-                LinearLayout {
-                    B()
-                }
-            }
-        }.then { activity ->
-
-            assertChildHierarchy(activity.root.firstViewBlockHolder()) {
-                """
-                <LinearLayout>
-                    <LinearLayout>
-                        <TextView />
-                    </LinearLayout>
-                    <LinearLayout>
-                        <TextView />
-                    </LinearLayout>
-                </LinearLayout>
-                """
-            }
-        }
-    }
-
-    @Test
-    @Ignore("b/179279455")
-    fun testCorrectViewTreeWithComponentWithMultipleRoots() {
-
-        @Composable fun B() {
-            TextView()
-            TextView()
-        }
-
-        compose {
-            LinearLayout {
-                LinearLayout {
-                    B()
-                }
-                LinearLayout {
-                    B()
-                }
-            }
-        }.then {
-
-            assertChildHierarchy(activity.root.firstViewBlockHolder()) {
-                """
-                <LinearLayout>
-                    <LinearLayout>
-                        <TextView />
-                        <TextView />
-                    </LinearLayout>
-                    <LinearLayout>
-                        <TextView />
-                        <TextView />
-                    </LinearLayout>
-                </LinearLayout>
-                """
-            }
-        }
-    }
-
-    @Test // regression b/157111271
-    fun testInsertDuringRecomposition() {
-        var includeA by mutableStateOf(false)
-        var someState by mutableStateOf(0)
-        var someOtherState by mutableStateOf(1)
-
-        @Composable fun B(@Suppress("UNUSED_PARAMETER") value: Int) {
-            // empty
-        }
-
-        @Composable fun A() {
-            B(someState)
-            someState++
-        }
-
-        @Composable fun T() {
-            subCompose {
-                // Take up some slot space
-                // This makes it more likely to reproduce bug 157111271.
-                remember(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15) {
-                    1
-                }
-                if (includeA) {
-                    Wrapper {
-                        B(0)
-                        B(someOtherState)
-                        B(2)
-                        B(3)
-                        B(4)
-                        A()
-                    }
-                }
-            }
-        }
-
-        compose {
-            T()
-        }.then {
-            includeA = true
-        }.then {
-            someOtherState = 10
-        }.then {
-            // force recompose
-        }
-    }
-
-    @Test // regression test for b/161892016
-    fun testMultipleRecompose() {
-        class A
-
-        var state1 by mutableStateOf(1)
-        var state2 by mutableStateOf(1)
-
-        @Composable fun validate(a: A?) {
-            assertNotNull(a)
-        }
-
-        @Composable fun use(@Suppress("UNUSED_PARAMETER") i: Int) { }
-
-        @Composable fun useA(a: A = A()) {
-            validate(a)
-            use(state2)
-        }
-
-        @Composable fun test() {
-            use(state1)
-            useA()
-        }
-
-        compose {
-            test()
-        }.then {
-            // Recompose test() skipping useA()
-            state1 = 2
-        }.then {
-            // Recompose useA(). In the bug this causes validate() to be passed a null because
-            // the recompose scope is updated with a lambda that captures the parameters when the
-            // default parameter expressions are skipped.
-            state2 = 2
-        }.then {
-            // force recompose to validate a.
-        }
     }
 
     @Test
@@ -501,102 +141,13 @@ class RecomposerTests : BaseComposeTest() {
     }
 }
 
-@Composable
-fun Wrapper(content: @Composable () -> Unit) {
-    content()
-}
-
-private fun View.firstViewBlockHolder(): ViewGroup = traversal()
-    .filterIsInstance<ViewGroup>()
-    // NOTE: Implementation dependence on Compose UI implementation detail
-    .first { it.javaClass.simpleName == "ViewBlockHolder" }
-
-private fun View.viewBlockHolders(): Sequence<ViewGroup> = traversal()
-    .filterIsInstance<ViewGroup>()
-    // NOTE: Implementation dependence on Compose UI implementation detail
-    .filter { it.javaClass.simpleName == "ViewBlockHolder" }
-
-fun assertChildHierarchy(root: ViewGroup, getHierarchy: () -> String) {
-    val realHierarchy = printChildHierarchy(root)
-
-    assertEquals(
-        normalizeString(getHierarchy()),
-        realHierarchy.trim()
-    )
-}
-
-fun assertChildHierarchy(roots: Sequence<ViewGroup>, getHierarchy: () -> String) {
-    val realHierarchy = roots.map { printChildHierarchy(it).trim() }.joinToString("\n")
-
-    assertEquals(
-        normalizeString(getHierarchy()),
-        realHierarchy.trim()
-    )
-}
-
-fun normalizeString(str: String): String {
-    val lines = str.split('\n').dropWhile { it.isBlank() }.dropLastWhile {
-        it.isBlank()
-    }
-    if (lines.isEmpty()) return ""
-    val toRemove = lines.first().takeWhile { it == ' ' }.length
-    return lines.joinToString("\n") { it.substring(Math.min(toRemove, it.length)) }
-}
-
-fun printChildHierarchy(root: ViewGroup): String {
-    val sb = StringBuilder()
-    for (i in 0 until root.childCount) {
-        printView(root.getChildAt(i), 0, sb)
-    }
-    return sb.toString()
-}
-
-fun printView(view: View, indent: Int, sb: StringBuilder) {
-    val whitespace = " ".repeat(indent)
-    val name = view.javaClass.simpleName
-    val attributes = printAttributes(view)
-    if (view is ViewGroup && view.childCount > 0) {
-        sb.appendLine("$whitespace<$name$attributes>")
-        for (i in 0 until view.childCount) {
-            printView(view.getChildAt(i), indent + 4, sb)
-        }
-        sb.appendLine("$whitespace</$name>")
-    } else {
-        sb.appendLine("$whitespace<$name$attributes />")
-    }
-}
-
-fun printAttributes(view: View): String {
-    val attrs = mutableListOf<String>()
-
-    // NOTE: right now we only look for id and text as attributes to print out... but we are
-    // free to add more if it makes sense
-    if (view.id != -1) {
-        attrs.add("id=${view.id}")
-    }
-
-    if (view is TextView && view.text.length > 0) {
-        attrs.add("text='${view.text}'")
-    }
-
-    val result = attrs.joinToString(" ", prefix = " ")
-    if (result.length == 1) {
-        return ""
-    }
-    return result
-}
-
 class Counter {
     private var counts = mutableMapOf<String, Int>()
     fun inc(key: String) = counts.getOrPut(key, { 0 }).let { counts[key] = it + 1 }
-    fun reset() {
-        counts = mutableMapOf()
-    }
-
     operator fun get(key: String) = counts[key] ?: 0
 }
 
-private class Trigger() {
+private class Trigger {
     val count = mutableStateOf(0)
     fun subscribe() { count.value }
     fun recompose() { count.value += 1 }
