+++ compose/runtime/runtime/src/test/kotlin/androidx/compose/runtime/RecomposerTests.kt
@@ -16,7 +16,11 @@
 
 package androidx.compose.runtime
 
+import androidx.compose.runtime.mock.Linear
 import androidx.compose.runtime.mock.TestMonotonicFrameClock
+import androidx.compose.runtime.mock.Text
+import androidx.compose.runtime.mock.compositionTest
+import androidx.compose.runtime.mock.expectNoChanges
 import androidx.compose.runtime.snapshots.Snapshot
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.Job
@@ -105,6 +109,139 @@ class RecomposerTests {
             )
         }
     }
+
+    @Test
+    fun testRecomposition() = compositionTest {
+        val counter = Counter()
+        val triggers = mapOf(
+            99 to Trigger(),
+            100 to Trigger(),
+            102 to Trigger(),
+        )
+        compose {
+            RecomposeTestComponentsA(
+                counter,
+                triggers
+            )
+        }
+
+        assertEquals(1, counter["A"])
+        assertEquals(1, counter["100"])
+        assertEquals(1, counter["101"])
+        assertEquals(1, counter["102"])
+
+        triggers[100]?.recompose()
+        triggers[102]?.recompose()
+
+        // nothing should happen synchronously
+        assertEquals(1, counter["A"])
+        assertEquals(1, counter["100"])
+        assertEquals(1, counter["101"])
+        assertEquals(1, counter["102"])
+
+        expectNoChanges()
+
+        assertEquals(1, counter["A"])
+        assertEquals(2, counter["100"])
+        assertEquals(1, counter["101"])
+        assertEquals(2, counter["102"])
+
+        // recompose() both the parent and the child... and show that the child only
+        // recomposes once as a result
+        triggers[99]?.recompose()
+        triggers[102]?.recompose()
+
+        expectNoChanges()
+
+        assertEquals(2, counter["A"])
+        assertEquals(3, counter["100"])
+        assertEquals(2, counter["101"])
+        assertEquals(3, counter["102"])
+    }
+
+    @Test // regression b/157111271
+    fun testInsertDuringRecomposition() = compositionTest {
+        var includeA by mutableStateOf(false)
+        var someState by mutableStateOf(0)
+        var someOtherState by mutableStateOf(1)
+
+        @Composable fun B(@Suppress("UNUSED_PARAMETER") value: Int) {
+            // empty
+        }
+
+        @Composable fun A() {
+            B(someState)
+            someState++
+        }
+
+        @Composable fun T() {
+            TestSubcomposition {
+                // Take up some slot space
+                // This makes it more likely to reproduce bug 157111271.
+                remember(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15) {
+                    1
+                }
+                if (includeA) {
+                    Wrapper {
+                        B(0)
+                        B(someOtherState)
+                        B(2)
+                        B(3)
+                        B(4)
+                        A()
+                    }
+                }
+            }
+        }
+
+        compose {
+            T()
+        }
+
+        includeA = true
+        advance(ignorePendingWork = true)
+
+        someOtherState = 10
+        advance(ignorePendingWork = true)
+
+        advance(ignorePendingWork = true)
+    }
+
+    @Test // regression test for b/161892016
+    fun testMultipleRecompose() = compositionTest {
+        class A
+
+        var state1 by mutableStateOf(1)
+        var state2 by mutableStateOf(1)
+
+        @Composable fun validate(a: A?) {
+            assertNotNull(a)
+        }
+
+        @Composable fun use(@Suppress("UNUSED_PARAMETER") i: Int) { }
+
+        @Composable fun useA(a: A = A()) {
+            validate(a)
+            use(state2)
+        }
+
+        @Composable fun test() {
+            use(state1)
+            useA()
+        }
+
+        compose {
+            test()
+        }
+
+        // Recompose test() skipping useA()
+        state1 = 2
+        advance()
+
+        state2 = 2
+        advance()
+        advance()
+    }
 }
 
 private class UnitApplier : Applier<Unit> {
@@ -131,4 +268,47 @@ private class UnitApplier : Applier<Unit> {
 
     override fun clear() {
     }
+}
+
+class Counter {
+    private var counts = mutableMapOf<String, Int>()
+    fun inc(key: String) = counts.getOrPut(key, { 0 }).let { counts[key] = it + 1 }
+    fun reset() {
+        counts = mutableMapOf()
+    }
+
+    operator fun get(key: String) = counts[key] ?: 0
+}
+
+@Composable
+private fun RecomposeTestComponentsA(counter: Counter, triggers: Map<Int, Trigger>) {
+    counter.inc("A")
+    triggers[99]?.subscribe()
+    Linear {
+        for (id in 100..102) {
+            key(id) {
+                RecomposeTestComponentsB(
+                    counter,
+                    triggers,
+                    id
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun RecomposeTestComponentsB(
+    counter: Counter,
+    triggers: Map<Int, Trigger>,
+    id: Int = 0
+) {
+    counter.inc("$id")
+    triggers[id]?.subscribe()
+    Text("$id")
+}
+
+@Composable
+private fun Wrapper(content: @Composable () -> Unit) {
+    content()
 }
